pyproject.toml
[tool.poetry]
name = "technezis-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lxml = "^5.3.1"
aiohttp = "^3.11.13"
selenium = "^4.29.0"
webdriver-manager = "^4.0.2"
flake8 = "^7.1.2"
python-telegram-bot = "^22.0"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
